Audio View: go to measure
Add measure controls to the audio player. Based on the midi information
the measure number is updated on playback. Entering upbeat duration and
measure number the user can set the start position for midi playback.

Midi meta events are analyzed for time and tempo changes. Start ticks
for each measure are based on asking the sequencer for the absolute
time. The measure controls are hidden if the midi analysis is not
possible.